+++ src/modules/mo-blog/enums/blog-permissions.enum.ts
@@ -1,5 +1,0 @@
-export enum BlogPermissionsEnum {
-  BLOG__ADMIN = 'BLOG__ADMIN',
-  BLOG__USER = 'BLOG__USER',
-  BLOG__VIEWER = 'BLOG__VIEWER'
-}