+++ src/modules/mo-changelog/data-transfer-objects/index.ts
@@ -1,3 +1,0 @@
-export * from './changelog.dto';
-export * from './changelog-create-request.dto';
-export * from './changelog-patch-request.dto';