+++ src/modules/mo-namespace/data-transfer-objects/namespace-service/namespace-service-pod/index.ts
@@ -1,2 +1,3 @@
 export * from './namespace-service-pod.dto';
+export * from './namespace-service-pod-flat.dto';
 export * from './namespace-service-pod-origin-traffic.dto';