+++ src/modules/mo-namespace/data-transfer-objects/namespace/namespace-flat.dto.ts
@@ -1,0 +1,35 @@
+import { Expose, Transform } from 'class-transformer';
+import { isArray, isString } from 'class-validator';
+import { BaseEntityDto } from '../../../mo-core';
+import { NamespaceStateEnum } from '../../enums';
+
+export class NamespaceFlatDto extends BaseEntityDto {
+  @Expose()
+  shortId: string;
+
+  @Transform(({ value, obj }) => (value && isString(value) && value.length > 0 ? value : obj.name))
+  @Expose()
+  displayName: string;
+
+  @Expose()
+  name: string;
+
+  @Transform(({ value }) => (value && isArray(value) ? value : []))
+  @Expose()
+  users: string[];
+
+  @Expose()
+  hostname: string;
+
+  @Expose()
+  domain: string;
+
+  @Expose()
+  description: string;
+
+  @Expose()
+  state: NamespaceStateEnum;
+
+  @Expose()
+  bgColorStyle: string;
+}