+++ src/modules/mo-notification/interfaces/notification-redis-pub-sub-message.ts
@@ -1,5 +1,5 @@
 import { NotificationSocketDataEventEnum, NotificationSocketEventEnum } from '../enums';
-import { IRedisPubSubMessage } from './redis-pub-sub-message';
+import { IRedisPubSubMessageDeprecated } from './redis-pub-sub-message';
 import { RedisPubSubMessageDataTypes } from '../types';
 
 /*********************************************
@@ -9,7 +9,7 @@ import { RedisPubSubMessageDataTypes } from '../types';
 /**
  * Hint: data type is NotificationSocketDataEventEnum
  */
-export interface IRedisPubSubMessageNotification extends IRedisPubSubMessage {
+export interface IRedisPubSubMessageNotification extends IRedisPubSubMessageDeprecated {
   redisChannel: 'mo_notification';
   socketEvent: NotificationSocketEventEnum.NOTIFICATION_SERVICE;
   data: RedisPubSubMessageDataTypes<NotificationSocketDataEventEnum>[];