+++ src/modules/mo-notification/types/redis-pub-sub-message.types.ts
@@ -1,19 +1,19 @@
 import {
-  IRedisPubSubMessageAuth,
-  IRedisPubSubMessageK8sNotification,
-  IRedisPubSubMessageNamespace,
-  IRedisPubSubMessageNamespaceInvitation,
-  IRedisPubSubMessageUser
+  // IRedisPubSubMessageAuth,
+  // IRedisPubSubMessageK8sNotification,
+  // IRedisPubSubMessageNamespace,
+  // IRedisPubSubMessageNamespaceInvitation,
+  // IRedisPubSubMessageUser
 } from '../interfaces';
 import { IRedisPubSubMessageNotification } from '../interfaces/notification-redis-pub-sub-message';
-import { IRedisPubSubMessagePayment } from '../interfaces/payment-redis-pub-sub-message';
+// import { IRedisPubSubMessagePayment } from '../interfaces/payment-redis-pub-sub-message';
 
 export type RedisPubSubMessageTypes =
   // | IRedisPubSubMessage
-  | IRedisPubSubMessageK8sNotification
-  | IRedisPubSubMessageNamespace
-  | IRedisPubSubMessageNamespaceInvitation
+  // | IRedisPubSubMessageK8sNotification
+  // | IRedisPubSubMessageNamespace
+  // | IRedisPubSubMessageNamespaceInvitation
   | IRedisPubSubMessageNotification
-  | IRedisPubSubMessageUser
-  | IRedisPubSubMessagePayment
-  | IRedisPubSubMessageAuth;
+  // | IRedisPubSubMessageUser
+  // | IRedisPubSubMessagePayment
+  // | IRedisPubSubMessageAuth;