+++ src/modules/mo-payment/product/data-transfer-objects/cluster/cluster-helm-templates.dto.ts
@@ -1,0 +1,24 @@
+import { Expose } from 'class-transformer';
+import { IsString } from 'class-validator';
+
+export class ClusterHelmTemplatesDto {
+  @IsString()
+  @Expose()
+  defaultBackendHelmStr: string;
+
+  @IsString()
+  @Expose()
+  certManagerHelmStr: string;
+
+  @IsString()
+  @Expose()
+  installHelmStr: string;
+
+  @IsString()
+  @Expose()
+  updateHelmStr: string;
+
+  @IsString()
+  @Expose()
+  deleteHelmStr: string;
+}