+++ src/modules/mo-payment/product/data-transfer-objects/cluster/cluster-patch-request.dto.ts
@@ -1,19 +1,7 @@
 import { ClusterCreateRequestDto } from './cluster-create-request.dto';
-import {
-  isArray,
-  IsBoolean,
-  isIP,
-  IsNotEmpty,
-  IsNumber,
-  IsOptional,
-  IsString,
-  IsUUID,
-  Matches,
-  MaxLength,
-  MinLength
-} from 'class-validator';
+import { isArray, IsBoolean, isIP, IsNotEmpty, IsNumber, IsOptional, IsString, IsUUID } from 'class-validator';
 import { Expose, plainToInstance, Transform, Type } from 'class-transformer';
-import { MoUtils } from '@mo/js-utils';
+import { IsSemanticVersion, MoUtils } from '@mo/js-utils';
 import { CountryDto } from '../../../../mo-core';
 import _ from 'lodash';
 import { ClusterSetupDto } from './cluster-setup.dto';
@@ -27,40 +15,13 @@ export class ClusterPatchRequestDto extends ClusterCreateRequestDto {
 
   @Type(() => CountryDto)
   @IsOptional()
+  @Transform(({ value }) => (value?.code ? value : undefined))
   @Expose()
   country: CountryDto;
 
   @IsOptional()
   @Expose()
   region: string;
-
-  @IsNotEmpty()
-  @IsString()
-  @Expose()
-  displayName: string;
-
-  @IsNotEmpty()
-  @IsString()
-  @MinLength(3)
-  @MaxLength(64)
-  @Matches(/^([a-z])([a-z0-9-_])/, {
-    message: '$property must conform to: a-z, 0-9, - ;min 3, max 64 char'
-  })
-  @Transform(({ value }) => {
-    if (!value) {
-      return 'mogenius';
-    }
-    return value
-      ?.toLowerCase()
-      .replace(/[^a-z0-9-]/g, '-')
-      ?.replace(/ +/g, '')
-      ?.substring(0, 64)
-      .split('-')
-      .filter((item: string) => item && item.length > 0)
-      .join('-');
-  })
-  @Expose()
-  name: string;
 
   @Transform(({ value }) =>
     _.uniq((value && isArray(value) ? value : []) as string[]).filter((item: string) => isIP(item))
@@ -103,7 +64,7 @@ export class ClusterPatchRequestDto extends ClusterCreateRequestDto {
   @IsOptional()
   @IsString()
   @Expose()
-  clusterId: string;
+  clusterMfaId: string;
 
   @Transform(({ value }) => MoUtils.parseBoolean(value))
   @IsOptional()
@@ -111,8 +72,35 @@ export class ClusterPatchRequestDto extends ClusterCreateRequestDto {
   @Expose()
   apiKeyIsActive: boolean;
 
+  @IsOptional()
+  @IsString()
+  @IsSemanticVersion()
+  @Expose()
+  appVersion: string
+
   @Type(() => ClusterSetupDto)
   @Transform(({ value }) => plainToInstance(ClusterSetupDto, value, { excludeExtraneousValues: true }))
   @Expose()
-  clusterSetup: string[];
+  clusterSetup: ClusterSetupDto;
+
+  @IsNotEmpty()
+  @IsString()
+  @Expose()
+  containerRegistryUrl: string;
+
+  @IsNotEmpty()
+  @Transform(({ value }) => MoUtils.removeLastSlashes(value))
+  @IsString()
+  @Expose()
+  containerRegistryPath: string;
+
+  @IsOptional()
+  @IsString()
+  @Expose()
+  containerRegistryUser: string;
+
+  @IsOptional()
+  @IsString()
+  @Expose()
+  containerRegistryPat: string;
 }