+++ src/modules/mo-payment/product/data-transfer-objects/cluster/cluster-resource-info-payload.dto.ts
@@ -1,0 +1,16 @@
+import { Expose, Transform, Type } from 'class-transformer';
+import { ClusterNodeDto } from './cluster-node.dto';
+import { BaseEntityDto } from '../../../../mo-core';
+import { isArray, IsString } from 'class-validator';
+import _ from 'lodash';
+
+export class ClusterResourceInfoPayloadDto extends BaseEntityDto {
+  @IsString({ each: true })
+  @Transform(({ value }) => (value && isArray(value) ? _.uniq(value) : []) as string[])
+  @Expose()
+  loadBalancerExternalIps: string[];
+
+  @Type(() => ClusterNodeDto)
+  @Expose()
+  nodeStats: ClusterNodeDto[];
+}