+++ src/modules/mo-payment/product/data-transfer-objects/cluster/current-k8s-manager-version.dto.ts
@@ -1,0 +1,8 @@
+import { Expose } from 'class-transformer';
+import { IsSemanticVersion } from '@mo/js-utils';
+
+export class CurrentK8sManagerVersionDto {
+  @IsSemanticVersion()
+  @Expose()
+  version: string;
+}