+++ src/modules/mo-sitemap/data-transfer-objects/index.ts
@@ -1,3 +1,0 @@
-export * from './sitemap.dto';
-export * from './sitemap-create-request.dto';
-export * from './sitemap-patch-request.dto';