+++ src/modules/mo-websocket/payment-redis-pub-sub-message.ts
@@ -1,0 +1,8 @@
+import { IRedisPubSubMessage, IRedisPubSubMessageData, WebsocketChannelEnum } from '@mo/websocket-dto';
+import { PaymentSocketDataEventEnum, PaymentSocketEventEnum } from '../mo-notification';
+
+export interface IRedisPubSubMessagePayment
+  extends IRedisPubSubMessage<PaymentSocketEventEnum, IRedisPubSubMessageData<PaymentSocketDataEventEnum, any>[]> {
+  redisChannel: WebsocketChannelEnum.MO_WEBSOCKET_PUBLISH_USER_MESSAGE;
+  socketEvent: PaymentSocketEventEnum.SUBSCRIPTION_SERVICE;
+}